Move AVLTree and BaseAVLTree into avl package
