+++ src/main/java/com/albion/common/tree/avl/BaseAVLTree.java
@@ -1,4 +1,4 @@
-package com.albion.common.tree;
+package com.albion.common.tree.avl;
 
 import com.albion.common.tree.node.TreeNode;
